prozorro/requirements.txt
requests>=2.31.0
pydantic>=2.5.0
openai>=1.0.0 
python-dotenv>=1.0.1
beautifulsoup4>=4.13.0
lxml>=5.0.0